+++ src/Handlers/GameUpdate.cs
@@ -23,6 +23,7 @@ namespace Amnesia.Handlers {
         }
 
         private static void HandleBloodMoon() {
+            // TODO: this should probably be injected into triggers when bloodmoon starts and ends, honestly
             try {
                 if (isBloodmoon == GameManager.Instance.World.aiDirector.BloodMoonComponent.BloodMoonActive) {
                     return;
@@ -32,12 +33,12 @@ namespace Amnesia.Handlers {
                 var players = GameManager.Instance.World.Players.list;
                 if (isBloodmoon) {
                     for (var i = 0; i < players.Count; i++) {
-                        _ = players[i].Buffs.AddBuff(Values.BloodmoonLifeProtectionBuff);
+                        _ = players[i].Buffs.AddBuff(Values.BuffBloodmoonLifeProtection);
                     }
                 } else {
                     for (var i = 0; i < players.Count; i++) {
-                        _ = players[i].Buffs.AddBuff(Values.PostBloodmoonLifeProtectionBuff);
-                        players[i].Buffs.RemoveBuff(Values.BloodmoonLifeProtectionBuff);
+                        _ = players[i].Buffs.AddBuff(Values.BuffPostBloodmoonLifeProtection);
+                        players[i].Buffs.RemoveBuff(Values.BuffBloodmoonLifeProtection);
                     }
                 }
             } catch (Exception e) {