+++ src/Handlers/PlayerSpawnedInWorld.cs
@@ -22,11 +22,13 @@ namespace Amnesia.Handlers {
                 switch (respawnType) {
                     case RespawnType.EnterMultiplayer: // first-time login for new player
                         _ = PlayerHelper.AddPositiveOutlookTime(player, Config.PositiveOutlookTimeOnFirstJoin);
+                        RefundHardenedMemory(clientInfo, player);
                         HandleStandardRespawnSteps(player);
                         break;
                     case RespawnType.JoinMultiplayer: // existing player rejoining
                         // grace period should continue only so long as you don't disconnect
-                        player.Buffs.RemoveBuff(Values.PostBloodmoonLifeProtectionBuff);
+                        player.Buffs.RemoveBuff(Values.BuffPostBloodmoonLifeProtection);
+                        RefundHardenedMemory(clientInfo, player);
                         HandleStandardRespawnSteps(player);
                         break;
                     case RespawnType.Died: // existing player returned from death
@@ -38,7 +40,7 @@ namespace Amnesia.Handlers {
                 log.Error("Failed to handle PlayerSpawnedInWorld event.", e);
             }
         }
-        
+
         /// <summary>
         /// Process steps common to enter/join/death.
         /// </summary>
@@ -47,27 +49,38 @@ namespace Amnesia.Handlers {
         private static void HandleStandardRespawnSteps(EntityPlayer player) {
 
             // Ensure joining/respawning players have their constants updated
-            if (player.GetCVar(Values.LongTermMemoryLevelCVar) != Config.LongTermMemoryLevel) {
-                player.SetCVar(Values.LongTermMemoryLevelCVar, Config.LongTermMemoryLevel);
+            if (player.GetCVar(Values.CVarLongTermMemoryLevel) != Config.LongTermMemoryLevel) {
+                player.SetCVar(Values.CVarLongTermMemoryLevel, Config.LongTermMemoryLevel);
             }
 
             // Remove Positive Outlook if admin disabled it since player's last login
-            if (Config.PositiveOutlookTimeOnMemoryLoss == 0 && player.Buffs.HasBuff(Values.PositiveOutlookBuff)) {
-                player.Buffs.RemoveBuff(Values.PositiveOutlookBuff);
+            if (Config.PositiveOutlookTimeOnMemoryLoss == 0 && player.Buffs.HasBuff(Values.BuffPositiveOutlook)) {
+                player.Buffs.RemoveBuff(Values.BuffPositiveOutlook);
             }
 
             // Apply/Remove memory protection based on configuration
             if (Config.ProtectMemoryDuringBloodmoon) {
                 // add or remove protection based on whether BM is active
                 if (GameManager.Instance.World.aiDirector.BloodMoonComponent.BloodMoonActive) {
-                    _ = player.Buffs.AddBuff(Values.BloodmoonLifeProtectionBuff);
+                    _ = player.Buffs.AddBuff(Values.BuffBloodmoonLifeProtection);
                 } else {
-                    player.Buffs.RemoveBuff(Values.BloodmoonLifeProtectionBuff);
+                    player.Buffs.RemoveBuff(Values.BuffBloodmoonLifeProtection);
                 }
             } else {
                 // remove/clean up since protection is inactive
-                player.Buffs.RemoveBuff(Values.BloodmoonLifeProtectionBuff);
-                player.Buffs.RemoveBuff(Values.PostBloodmoonLifeProtectionBuff);
+                player.Buffs.RemoveBuff(Values.BuffBloodmoonLifeProtection);
+                player.Buffs.RemoveBuff(Values.BuffPostBloodmoonLifeProtection);
+            }
+        }
+
+        /// <summary>
+        /// Temporary method to automatically refund any players with the Hardened Memory buff from version 1.0.0.
+        /// </summary>
+        /// <param name="player">EntityPlayer to check buffs for and refund if hardened.</param>
+        private static void RefundHardenedMemory(ClientInfo clientInfo, EntityPlayer player) {
+            if (player.Buffs.HasBuff(Values.BuffHardenedMemory)) {
+                PlayerHelper.GiveItem(clientInfo, player, Values.NameMemoryBoosters);
+                player.Buffs.RemoveBuff(Values.BuffHardenedMemory);
             }
         }
     }