+++ src/Utilities/PlayerHelper.cs
@@ -33,10 +33,6 @@ namespace Amnesia.Utilities {
                 player.Progression.ExpToNextLevel = player.Progression.GetExpForNextLevel();
                 player.Progression.SkillPoints = Config.LongTermMemoryLevel - 1;
 
-                // Zero out xp debt; the reset has caused enough suffering ;)
-                player.Progression.ExpDeficit = 0;
-                player.SetCVar("_expdeficit", 0);
-
                 // Return all skill points rewarded from completed quest; should cover vanilla quest_BasicSurvival8, for example
                 if (!Config.ForgetIntroQuests) {
                     try {
@@ -76,11 +72,11 @@ namespace Amnesia.Utilities {
                 ConnectionManager.Instance.SendPackage(NetPackageManager.GetPackage<NetPackagePlayerStats>().Setup(player), false, player.entityId);
             }
 
-            _ = player.Buffs.AddBuff(Values.MemoryLossNotificationBuff);
+            _ = player.Buffs.AddBuff(Values.BuffMemoryLoss);
             if (Config.PositiveOutlookTimeOnMemoryLoss > 0) {
                 log.Trace($"{player.GetDebugName()} will receive the Positive Outlook buff.");
-                player.SetCVar(Values.PositiveOutlookRemTimeCVar, Config.PositiveOutlookTimeOnMemoryLoss);
-                _ = player.Buffs.AddBuff(Values.PositiveOutlookBuff);
+                player.SetCVar(Values.CVarPositiveOutlookRemTime, Config.PositiveOutlookTimeOnMemoryLoss);
+                _ = player.Buffs.AddBuff(Values.BuffPositiveOutlook);
             }
         }
 
@@ -88,11 +84,11 @@ namespace Amnesia.Utilities {
             if (valueToAdd == 0) {
                 return 0;
             }
-            var playerRemTime = Math.Max(0, player.GetCVar(Values.PositiveOutlookRemTimeCVar));
+            var playerRemTime = Math.Max(0, player.GetCVar(Values.CVarPositiveOutlookRemTime));
             var targetValue = Math.Min(playerRemTime + valueToAdd, Config.PositiveOutlookMaxTime);
-            player.SetCVar(Values.PositiveOutlookRemTimeCVar, targetValue);
-            if (!player.Buffs.HasBuff(Values.PositiveOutlookBuff)) {
-                _ = player.Buffs.AddBuff(Values.PositiveOutlookBuff);
+            player.SetCVar(Values.CVarPositiveOutlookRemTime, targetValue);
+            if (!player.Buffs.HasBuff(Values.BuffPositiveOutlook)) {
+                _ = player.Buffs.AddBuff(Values.BuffPositiveOutlook);
             }
             return targetValue;
         }
@@ -106,6 +102,18 @@ namespace Amnesia.Utilities {
         public static void GiveItem(EntityPlayer player, string itemName, int count = 1) {
             var itemStack = new ItemStack(ItemClass.GetItem(itemName, true), count);
             var clientInfo = ConnectionManager.Instance.Clients.ForEntityId(player.entityId);
+            GiveItemStack(clientInfo, player.GetBlockPosition(), itemStack);
+        }
+
+        /// <summary>
+        /// Give an item to the player, placing it in the player's inventory if possible.
+        /// </summary>
+        /// <param name="clientInfo">ClientInfo for player to send the network package to.</param>
+        /// <param name="player">EntityPlayer to give item to.</param>
+        /// <param name="itemName">Name of the item to give the player.</param>
+        /// <param name="count">Number of items to give within a single stack (only works with stackable items).</param>
+        public static void GiveItem(ClientInfo clientInfo, EntityPlayer player, string itemName, int count = 1) {
+            var itemStack = new ItemStack(ItemClass.GetItem(itemName, true), count);
             GiveItemStack(clientInfo, player.GetBlockPosition(), itemStack);
         }
 
@@ -123,7 +131,5 @@ namespace Amnesia.Utilities {
             clientInfo.SendPackage(NetPackageManager.GetPackage<NetPackageEntityCollect>().Setup(entityId, clientInfo.entityId));
             _ = GameManager.Instance.World.RemoveEntity(entityId, EnumRemoveEntityReason.Despawned);
         }
-
-        internal static void GiveItem(EntityPlayer player, object memoryBoosterItemName) => throw new NotImplementedException();
     }
 }